Add add-remove.js and index.js file
